feat(config): moved conf.py into settings
