+++ backend/app/utils/data/data_source.py
@@ -7,7 +7,7 @@
 """
 
 from psycopg2 import connect
-from app.utils.config.conf import settings
+from app.utils.settings.config import settings
 from app.utils.data.queries import (
     QUERY_CREATE_TABLE,
     QUERY_CREATE_INDEX,
