+++ backend/app/utils/settings/__init__.py
@@ -1,0 +1,7 @@
+"""
+
+ Copyright 2024 @Qreater
+ Licensed under the Apache License, Version 2.0.
+ See: http://www.apache.org/licenses/LICENSE-2.0
+
+"""
